Updated C++ support for containers.
The compiler was changed to generate a richer C++ interface for container fields in data structures.  Methods are now available to query the container size and to work with individual elements.  Vectors can also be resized.

A pair of bugs associated with sets in the C++ library were also fixed.