Refactor serialization buffer sizing for Fortran
Before serializing data, the C++ and Fortran libraries must allocate a buffer large enough to hold the encoded bytes.  The previous method did not work for nested containers.  Here, the Fortran method is refactored to try to eliminate these problems.  Serialization for C++ had been updated in a previous commit.